add playback store module
