+++ client/store/playback/mutations.ts
@@ -1,0 +1,172 @@
+/* eslint-disable no-param-reassign */
+import { Mutations } from 'typed-vuex';
+import { PlaybackState } from './state';
+import type { SpotifyAPI, ZeroToHundred } from '~~/types';
+
+import { convertUriToId } from '~/scripts/converter/convertUriToId';
+
+export type PlaybackMutations = {
+  SET_DEVICE_ID: string | undefined
+  SET_ACTIVE_DEVICE_ID: string | undefined
+  SET_DEVICE_LIST: SpotifyAPI.Device[]
+  SET_CUSTOM_CONTEXT_URI: string | undefined
+  SET_CUSTOM_TRACK_URI_LIST: string[] | undefined
+  SET_RECENTLY_PLAYED: SpotifyAPI.Player.RecentlyPlayed | undefined
+  SET_GET_CURRENT_PLAYBACK_TIMER_ID: ReturnType<typeof setTimeout> | number | undefined
+  INCREMENT_RETRY_COUNTS_OF_GET_CURRENT_PLAYBACK: void
+  RESET_RETRY_COUNTS_OF_GET_CURRENT_PLAYBACK: void
+  SET_CURRENT_TRACK: Spotify.Track | undefined
+  SET_NEXT_TRACK_LIST: Spotify.Track[]
+  SET_PREVIOUS_TRACK_LIST: Spotify.Track[]
+  SET_IS_SAVED_TRACK: boolean
+  SET_IS_PLAYING: boolean
+  SET_CONTEXT_URI: string | undefined
+  SET_POSITION_MS: number
+  SET_DISABLED_PLAYING_FROM_BEGINING: boolean
+  SET_DURATION_MS: number
+  SET_IS_SHUFFLED: boolean
+  SET_REPEAT_MODE: 0 | 1 | 2
+  SET_DISALLOWS: SpotifyAPI.Disallows
+  SET_VOLUME_PERCENT: { volumePercent: ZeroToHundred }
+  SET_IS_MUTED: boolean
+};
+
+export type RootMutations = {
+  ['playback/SET_DEVICE_ID']: PlaybackMutations['SET_DEVICE_ID']
+  ['playback/SET_ACTIVE_DEVICE_ID']: PlaybackMutations['SET_ACTIVE_DEVICE_ID']
+  ['playback/SET_DEVICE_LIST']: PlaybackMutations['SET_DEVICE_LIST']
+  ['playback/SET_CUSTOM_CONTEXT_URI']: PlaybackMutations['SET_CUSTOM_CONTEXT_URI']
+  ['playback/SET_CUSTOM_TRACK_URI_LIST']: PlaybackMutations['SET_CUSTOM_TRACK_URI_LIST']
+  ['playback/SET_RECENTLY_PLAYED']: PlaybackMutations['SET_RECENTLY_PLAYED']
+  ['playback/SET_GET_CURRENT_PLAYBACK_TIMER_ID']: PlaybackMutations['SET_GET_CURRENT_PLAYBACK_TIMER_ID']
+  ['playback/INCREMENT_RETRY_COUNTS_OF_GET_CURRENT_PLAYBACK']: PlaybackMutations['INCREMENT_RETRY_COUNTS_OF_GET_CURRENT_PLAYBACK']
+  ['playback/RESET_RETRY_COUNTS_OF_GET_CURRENT_PLAYBACK']: PlaybackMutations['RESET_RETRY_COUNTS_OF_GET_CURRENT_PLAYBACK']
+  ['playback/SET_CURRENT_TRACK']: PlaybackMutations['SET_CURRENT_TRACK']
+  ['playback/SET_NEXT_TRACK_LIST']: PlaybackMutations['SET_NEXT_TRACK_LIST']
+  ['playback/SET_PREVIOUS_TRACK_LIST']: PlaybackMutations['SET_PREVIOUS_TRACK_LIST']
+  ['playback/SET_IS_SAVED_TRACK']: PlaybackMutations['SET_IS_SAVED_TRACK']
+  ['playback/SET_IS_PLAYING']: PlaybackMutations['SET_IS_PLAYING']
+  ['playback/SET_CONTEXT_URI']: PlaybackMutations['SET_CONTEXT_URI']
+  ['playback/SET_POSITION_MS']: PlaybackMutations['SET_POSITION_MS']
+  ['playback/SET_DISABLED_PLAYING_FROM_BEGINING']: PlaybackMutations['SET_DISABLED_PLAYING_FROM_BEGINING']
+  ['playback/SET_DURATION_MS']: PlaybackMutations['SET_DURATION_MS']
+  ['playback/SET_IS_SHUFFLED']: PlaybackMutations['SET_IS_SHUFFLED']
+  ['playback/SET_REPEAT_MODE']: PlaybackMutations['SET_REPEAT_MODE']
+  ['playback/SET_DISALLOWS']: PlaybackMutations['SET_DISALLOWS']
+  ['playback/SET_VOLUME_PERCENT']: PlaybackMutations['SET_VOLUME_PERCENT']
+  ['playback/SET_IS_MUTED']: PlaybackMutations['SET_IS_MUTED']
+};
+
+const mutations: Mutations<PlaybackState, PlaybackMutations> = {
+  SET_DEVICE_ID(state, deviceId) {
+    state.deviceId = deviceId;
+  },
+
+  SET_ACTIVE_DEVICE_ID(state, activeDeviceId) {
+    state.activeDeviceId = activeDeviceId;
+  },
+
+  SET_DEVICE_LIST(state, deviceList) {
+    state.deviceList = deviceList;
+  },
+
+  SET_CUSTOM_CONTEXT_URI(state, contextUri) {
+    state.customContextUri = contextUri;
+  },
+
+  SET_CUSTOM_TRACK_URI_LIST(state, trackUriList) {
+    state.customTrackUriList = trackUriList;
+  },
+
+  SET_RECENTLY_PLAYED(state, recentlyPlayed) {
+    state.recentlyPlayed = recentlyPlayed;
+  },
+
+  SET_GET_CURRENT_PLAYBACK_TIMER_ID(state, timer) {
+    const { getCurrentPlaybackTimer } = state;
+    if (typeof getCurrentPlaybackTimer === 'number') {
+      // クライアントサイドで実行
+      window.clearTimeout(getCurrentPlaybackTimer);
+    } else if (getCurrentPlaybackTimer != null) {
+      // サーバーサイドで実行
+      clearTimeout(getCurrentPlaybackTimer);
+    }
+
+    state.getCurrentPlaybackTimer = timer;
+  },
+  INCREMENT_RETRY_COUNTS_OF_GET_CURRENT_PLAYBACK(state) {
+    state.retryCountsOfGetCurrentPlayback += 1;
+  },
+
+  RESET_RETRY_COUNTS_OF_GET_CURRENT_PLAYBACK(state) {
+    state.retryCountsOfGetCurrentPlayback = 0;
+  },
+
+  SET_CURRENT_TRACK(state, currentTrack) {
+    state.artworkList = currentTrack?.album.images;
+    state.trackName = currentTrack?.name;
+    state.trackId = currentTrack?.id ?? undefined;
+    state.trackUri = currentTrack?.uri;
+    state.releaseName = currentTrack?.album.name;
+    state.releaseUri = currentTrack?.album.uri;
+    state.artistList = currentTrack?.artists.map((artist) => ({
+      name: artist.name,
+      id: convertUriToId(artist.uri),
+    }));
+  },
+
+  SET_NEXT_TRACK_LIST(state, nextTrackList) {
+    state.nextTrackList = nextTrackList;
+  },
+
+  SET_PREVIOUS_TRACK_LIST(state, previousTrackList) {
+    state.previousTrackList = previousTrackList;
+  },
+
+  SET_IS_SAVED_TRACK(state, isSavedTrack) {
+    state.isSavedTrack = isSavedTrack;
+  },
+
+  SET_IS_PLAYING(state, isPlaying) {
+    state.isPlaying = isPlaying;
+  },
+
+  SET_CONTEXT_URI(state, uri) {
+    state.contextUri = uri;
+  },
+
+  SET_POSITION_MS(state, positionMs) {
+    state.positionMs = positionMs;
+    state.disabledPlayingFromBegining = positionMs <= 1000;
+  },
+
+  SET_DISABLED_PLAYING_FROM_BEGINING(state, disabledPlayingFromBegining) {
+    state.disabledPlayingFromBegining = disabledPlayingFromBegining;
+  },
+
+  SET_DURATION_MS(state, durationMs) {
+    state.durationMs = durationMs;
+  },
+
+  SET_IS_SHUFFLED(state, isShuffled) {
+    state.isShuffled = isShuffled;
+  },
+
+  SET_REPEAT_MODE(state, repeatMode) {
+    state.repeatMode = repeatMode;
+  },
+
+  SET_DISALLOWS(state, disallows) {
+    state.disallows = disallows;
+  },
+
+  SET_VOLUME_PERCENT(state, { volumePercent }) {
+    state.volumePercent = volumePercent;
+  },
+
+  SET_IS_MUTED(state, isMuted) {
+    state.isMuted = isMuted;
+  },
+};
+
+export default mutations;
